+++ HopStepSnake/HopStepSnake/Snake.cpp
@@ -14,8 +14,8 @@ Snake::Snake()
 	for (int i = 0; i < 4; ++i)
 	{
 		HopStep::HSRect* body = new HopStep::HSRect(primitiveRect);
-		body->m_Center.x = x * 20 + 10;
-		body->m_Center.y = y * 20 + 10;
+		body->m_Center.x = static_cast<float>(x * 20 + 10);
+		body->m_Center.y = static_cast<float>(y * 20 + 10);
 		bodys.push_back(body);
 		body->SetVisible(true);
 		--x;
@@ -81,8 +81,8 @@ bool Snake::IsOverlappedPosition(int x, int y)
 HopStep::HSVector<int> Snake::GetHeadPosition() const
 {
 	HopStep::HSVector<int> pos(0);
-	pos.x = bodys[0]->m_Center.x;
-	pos.y = bodys[0]->m_Center.y;
+	pos.x = static_cast<int>(bodys[0]->m_Center.x);
+	pos.y = static_cast<int>(bodys[0]->m_Center.y);
 	return pos;
 }
